Models fix, ProducesResponseType fix

+++ EcoMonitor/Model/DTO/EnvFactorDTO.cs
@@ -17,7 +17,7 @@ namespace EcoMonitor.Model.DTO
         public double factor_Ca_value { get; set; }
         [Range(0, double.MaxValue)]
         public double factor_Ch_value { get; set; }
-        [Range(0, double.MaxValue)]
+        [Range(0, int.MaxValue)]
         public int? rfc_factor_id { get; set; }
 
     }
